+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
 pub mod data;
+pub mod error;
 pub mod model;
 pub mod threadpool;
